[executable-helpers] cleanup and standardize crate name
