+++ admission_control/admission-control-service/Cargo.toml
@@ -19,7 +19,7 @@ config = { path = "../../config" }
 crypto = { path = "../../crypto/crypto" }
 debug_interface = { path = "../../common/debug_interface" }
 failure = { package = "failure_ext", path = "../../common/failure_ext" }
-executable_helpers = { path = "../../common/executable_helpers" }
+executable-helpers = { path = "../../common/executable-helpers" }
 grpc_helpers = { path = "../../common/grpc_helpers" }
 logger = { path = "../../common/logger" }
 mempool = { path = "../../mempool" }
