+++ crypto/secret-service/Cargo.toml
@@ -19,7 +19,7 @@ config = { path = "../../config" }
 crypto = { path = "../crypto" }
 crypto_derive = { path = "../crypto_derive" }
 debug_interface = { path = "../../common/debug_interface" }
-executable_helpers = { path = "../../common/executable_helpers" }
+executable-helpers = { path = "../../common/executable-helpers" }
 failure = { package = "failure_ext", path = "../../common/failure_ext" }
 grpc_helpers = { path = "../../common/grpc_helpers" }
 logger = { path = "../../common/logger" }
