+++ libra-node/Cargo.toml
@@ -22,7 +22,7 @@ config = { path = "../config" }
 consensus = { path = "../consensus" }
 crash_handler = { path = "../common/crash_handler" }
 debug_interface = { path = "../common/debug_interface" }
-executable_helpers = { path = "../common/executable_helpers" }
+executable-helpers = { path = "../common/executable-helpers" }
 executor = { path = "../execution/executor" }
 futures = { version = "=0.3.0-alpha.19", package = "futures-preview", features = ["async-await", "io-compat", "compat"] }
 grpc_helpers = { path = "../common/grpc_helpers" }
