+++ storage/storage-service/src/main.rs
@@ -3,7 +3,7 @@
 
 use config::config::NodeConfig;
 use debug_interface::{node_debug_service::NodeDebugService, proto::node_debug_interface_grpc};
-use executable_helpers::helpers::setup_executable_new;
+use executable_helpers::helpers::setup_executable;
 use failure::prelude::*;
 use grpc_helpers::spawn_service_thread;
 use logger::prelude::*;
@@ -62,7 +62,7 @@ fn main() {
     let args = Args::from_args();
 
     let (config, _logger) =
-        setup_executable_new(args.config.as_ref().map(PathBuf::as_path), args.no_logging);
+        setup_executable(args.config.as_ref().map(PathBuf::as_path), args.no_logging);
 
     let storage_node = StorageNode::new(config);
